Added Auth checking for users index (if is_admin).

+++ resources/views/users/table.blade.php
@@ -4,10 +4,15 @@
         <th>User Name</th>
         <th>First Name</th>
         <th>Last Name</th>
-        <th>Email</th>
-        <th>Password</th>
-        <th>Bitcoin Address</th>
-        <th>Is Admin</th>
+
+        @if(Auth::user() != null)
+            @if(Auth::user()->is_admin == true)
+                <th>Email</th>
+                <th>Password</th>
+                <th>Bitcoin Address</th>
+                <th>Is Admin</th>
+            @endif
+        @endif
         <th>Slug</th>
         <th colspan="3">Action</th>
     </thead>
@@ -17,19 +22,27 @@
             <td>{!! $user->user_name !!}</td>
             <td>{!! $user->first_name !!}</td>
             <td>{!! $user->last_name !!}</td>
-            <td>{!! $user->email !!}</td>
-            <td> ************ </td>
-            <td>{!! $user->bitcoin_address !!}</td>
-            <td>{!! $user->isAnAdmin() !!}</td>
+            @if(Auth::user() != null)
+                @if(Auth::user()->is_admin == true)
+                    <td>{!! $user->email !!}</td>
+                    <td> ************ </td>
+                    <td>{!! $user->bitcoin_address !!}</td>
+                    <td>{!! $user->isAnAdmin() !!}</td>
+                @endif
+            @endif
             <td>{!! $user->slug !!}</td>
             <td>
-                {!! Form::open(['route' => ['users.destroy', $user->slug], 'method' => 'delete']) !!}
                 <div class='btn-group'>
                     <a href="{!! route('users.show', [$user->slug]) !!}" class='btn btn-default btn-xs'><i class="glyphicon glyphicon-eye-open"></i></a>
-                    <a href="{!! route('users.edit', [$user->slug]) !!}" class='btn btn-default btn-xs'><i class="glyphicon glyphicon-edit"></i></a>
-                    {!! Form::button('<i class="glyphicon glyphicon-trash"></i>', ['type' => 'submit', 'class' => 'btn btn-danger btn-xs', 'onclick' => "return confirm('Are you sure?')"]) !!}
+                    @if(Auth::user() != null)
+                        @if(Auth::user()->is_admin == true)
+                            <a href="{!! route('users.edit', [$user->slug]) !!}" class='btn btn-default btn-xs'><i class="glyphicon glyphicon-edit"></i></a>
+                            {!! Form::open(['route' => ['users.destroy', $user->slug], 'method' => 'delete']) !!}
+                            {!! Form::button('<i class="glyphicon glyphicon-trash"></i>', ['type' => 'submit', 'class' => 'btn btn-danger btn-xs', 'onclick' => "return confirm('Are you sure?')"]) !!}
+                            {!! Form::close() !!}
+                        @endif
+                    @endif
                 </div>
-                {!! Form::close() !!}
             </td>
         </tr>
     @endforeach
